requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
isort==4.3.21
jose==1.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
